agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "endpointclaw"
version = "0.1.0"
description = "EndpointClaw — Local AI agent for endpoint monitoring and assistance"
readme = "README.md"
license = "Proprietary"
requires-python = ">=3.10"
authors = [
    { name = "Corvex Roofing Solutions" },
]

dependencies = [
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "watchdog>=4.0.0",
    "pystray>=0.19.0",
    "Pillow>=10.0.0",
    "PyMuPDF>=1.24.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "psutil>=5.9.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
endpointclaw = "agent.src.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
